update main and components
