+++ client/src/Components/Nav/nav.js
@@ -1,0 +1,26 @@
+// import React, { useState } from 'react';
+import React from 'react';
+
+function Nav(props) {
+    return (
+        <nav class='topnav'>
+            <ul className='flex-row'>
+                <li className="mx-2">
+                    <button onClick={() => props.setCurrentPage("SignIn")}>
+                        <a>Sign In</a>
+                    </button>
+                </li>
+
+                <li className='mx-2'>
+                    <button onClick={() => props.setCurrentPage ("SignUp")}>
+                        Sign Up
+                    </button>
+                </li>
+
+            </ul>
+        </nav>
+
+    )
+}
+
+export default Nav;
